refactor: split functions into their own files
